Fix issue with test failing due to unnecessary mocking.
