+++ ksql-rest-app/src/test/java/io/confluent/ksql/rest/server/execution/ListQueriesExecutorTest.java
@@ -23,7 +23,9 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 import com.google.common.collect.ImmutableList;
+import com.google.common.collect.ImmutableSet;
 import io.confluent.ksql.engine.KsqlEngine;
+import io.confluent.ksql.query.QueryId;
 import io.confluent.ksql.rest.entity.EntityQueryId;
 import io.confluent.ksql.rest.entity.Queries;
 import io.confluent.ksql.rest.entity.QueryDescription;
@@ -58,7 +60,7 @@ public class ListQueriesExecutorTest {
   public void shouldListQueriesBasic() {
     // Given
     final ConfiguredStatement<?> showQueries = engine.configure("SHOW QUERIES;");
-    final PersistentQueryMetadata metadata = TemporaryEngine.givenPersistentQuery("id");
+    final PersistentQueryMetadata metadata = givenPersistentQuery("id");
 
     final KsqlEngine engine = mock(KsqlEngine.class);
     when(engine.getPersistentQueries()).thenReturn(ImmutableList.of(metadata));
@@ -81,7 +83,7 @@ public class ListQueriesExecutorTest {
   public void shouldListQueriesExtended() {
     // Given
     final ConfiguredStatement<?> showQueries = engine.configure("SHOW QUERIES EXTENDED;");
-    final PersistentQueryMetadata metadata = TemporaryEngine.givenPersistentQuery("id");
+    final PersistentQueryMetadata metadata = givenPersistentQuery("id");
 
     final KsqlEngine engine = mock(KsqlEngine.class);
     when(engine.getPersistentQueries()).thenReturn(ImmutableList.of(metadata));
@@ -97,4 +99,13 @@ public class ListQueriesExecutorTest {
         QueryDescription.forQueryMetadata(metadata)));
   }
 
+  @SuppressWarnings("SameParameterValue")
+  public static PersistentQueryMetadata givenPersistentQuery(final String id) {
+    final PersistentQueryMetadata metadata = mock(PersistentQueryMetadata.class);
+    when(metadata.getQueryId()).thenReturn(new QueryId(id));
+    when(metadata.getSinkNames()).thenReturn(ImmutableSet.of(id));
+    when(metadata.getResultSchema()).thenReturn(TemporaryEngine.SCHEMA);
+
+    return metadata;
+  }
 }
